# Conflicts:
#	controller/PlongeeController.php
#	controller/PlongeurController.php
#	model/manager/SiteManager.php
#	view/plongeur/plongeur_show/plongeur_editform.html

+++ controller/PlongeurController.php
@@ -1,6 +1,7 @@
 <?php
 
 require_once('_ControllerClass.php');
+require_once('/model/utils/traitement.php');
 
 class PlongeurController extends _ControllerClass
 {
@@ -21,6 +22,10 @@ class PlongeurController extends _ControllerClass
         if ($urlSize > 1)
             if($url[1] == 'show')
                 $this->show();
+            if($url[1] == 'edit')
+                $this->edit();
+            else if($url[1] == 'delete')
+                $this->delete();
             else
                 throw new Exception('Page introuvable');
     }
@@ -44,18 +49,18 @@ class PlongeurController extends _ControllerClass
             if (!empty($_POST['searchPrenom']))
                 $search['prenom'] = $_POST['searchPrenom'];
 
-            if (!isset($_POST['searchInactive']))
+            if (isset($_POST['searchInactive']))
                 $search['inactive'] = '';
 
-            if (!empty($_POST['searchNom']) || !empty($_POST['searchPrenom']) || !isset($_POST['searchInactive']))
+            if (!empty($_POST['searchNom']) || !empty($_POST['searchPrenom']) || isset($_POST['searchInactive']))
                 $searchedPlongeurs = $this->plongeurManager->getSearchResult($search);
 
         }
 
-       // $this->delete();
+
 
         (new View('plongeur/plongeur_index'))->generate([
-            'allPlongeurs' => $this->plongeurManager->getAll(),
+            'allPlongeurs' => $this->plongeurManager->getAllActive(),
             'allAptitudes' => $this->aptitudeManager->getAll(),
             'searchedPlongeurs' => $searchedPlongeurs
         ]);
@@ -94,18 +99,22 @@ class PlongeurController extends _ControllerClass
         if ( isset($_POST['nom']) && isset($_POST['prenom']) && isset($_POST['aptitude']) ) {
 
             if (!empty($_POST['nom']) && !empty($_POST['prenom']) && !empty($_POST['aptitude'])) {
-                $nom = strtoupper($_POST['nom']);
-                $prenom = ucfirst($_POST['prenom']);
+                $nom = $_POST['nom'];
+                $prenom = $_POST['prenom'];
+                if($this->verifEntree($nom,$prenom)) {
+                    $plongeur[] = new Plongeur([]);
+                    $personne[] = new Personne([
+                        'PER_NOM' => traitementNom($nom),
+                        'PER_PRENOM' => traitementPrenom($prenom)
+                    ]);
 
-                $plongeur[] = new Plongeur([]);
-                $personne[] = new Personne([
-                    'PER_NOM' => $nom,
-                    'PER_PRENOM' => $prenom
-                ]);
+                    $plongeur[0]->setPersonne($personne);
 
-                $plongeur[0]->setPersonne($personne);
+                    $this->verification($plongeur, true);
+                }
+                else
+                    echo "le nom ou le prénom n'est pas correct";
 
-                $this->verification($plongeur, true);
 
             } else
                 echo 'Tous les champs ne sont pas remplis.';
@@ -121,9 +130,25 @@ class PlongeurController extends _ControllerClass
             'PER_NUM' => $_GET['per_num']]);
 
         if (is_null($plongeur))
-            header('location: plongeur');
+            header('location: /plongeur');
 
-        $this->plongeurManager->delete($plongeur);
+        if ( isset($_POST['submit']) ) {
+            $this->plongeurManager->delete($plongeur);
+            header('location: /plongeur');
+        }
+
+        (new View('plongeur/plongeur_removeform'))->generate([
+            'plongeur' => $plongeur,
+
+        ]);
+
+    }
+
+    public function verifEntree($nom,$prenom){
+        if(prenomCorrect($prenom) && nomCorrect($nom))
+            return true;
+        else
+            return false;
     }
 
     private function verification($plongeur, $add = false)
@@ -131,9 +156,10 @@ class PlongeurController extends _ControllerClass
         if ( $add || (isset($_POST['nom']) && isset($_POST['prenom']) && isset($_POST['aptitude'])) ) {
 
                 if ($add || (!empty($_POST['nom']) && !empty($_POST['prenom']) && !empty($_POST['aptitude']))) {
-                    $nom = strtoupper($_POST['nom']);
-                    $prenom = ucfirst($_POST['prenom']);
+                    $nom = $_POST['nom'];
+                    $prenom = $_POST['prenom'];
                     $aptitude = $_POST['aptitude'];
+
 
 
                     $personnes = $this->personneManager->getAll();
@@ -149,19 +175,23 @@ class PlongeurController extends _ControllerClass
                         $i = $nbPersonnes;
 
                     if ($i == $nbPersonnes) {
-                        $plongeur[0]->getPersonne()[0]->setPerNom($nom);
-                        $plongeur[0]->getPersonne()[0]->setPerPrenom($prenom);
+                        if($this->verifEntree($nom,$prenom)) {
+                            $plongeur[0]->getPersonne()[0]->setPerNom($nom);
+                            $plongeur[0]->getPersonne()[0]->setPerPrenom($prenom);
 
-                        $aptitudeObject = $this->aptitudeManager->getOne(['APT_CODE' => $aptitude]);
+                            $aptitudeObject = $this->aptitudeManager->getOne(['APT_CODE' => $aptitude]);
 
-                        if (!is_null($aptitudeObject))
-                            if ($add)
-                                $plongeur[0]->setAptitude($aptitudeObject);
+                            if (!is_null($aptitudeObject))
+                                if ($add)
+                                    $plongeur[0]->setAptitude($aptitudeObject);
                             $plongeur[0]->setAptCode($aptitude);
 
-                        $this->plongeurManager->update($plongeur, $add);
+                            $this->plongeurManager->update($plongeur, $add);
+                            header('location: '.URL.'/plongeur');
+                        }
+                        else
+                                echo "le nom ou le prénom n'est pas correct";
 
-                        header('location: '.URL.'/plongeur');
                     } else
                         echo 'Personne déjà enregistrée.';
                 } else
